chore: SRP(단일 책임 원칙)과 명료한 네이밍을 위한 구조 변경
