+++ dev/train/trainData/mnist/mnistTestData.js
