+++ dev/train/trainData/mnist/mnistTrainData.js
