+++ src/core/ops/activationOps.js
